Relax cooking time constraint to only apply if meal prep time is above a threshold

+++ database/database_tools.go
@@ -21,7 +21,7 @@ func RunMe() {
 type Meal struct {
 	ID          int
 	MealName    string
-	CookingTime float32
+	CookingTime float64
 	Category    string
 	LunchOnly   bool
 }
